+++ packages/livechat/src/lib/triggers.js
@@ -6,6 +6,13 @@ import { actions } from './triggerActions';
 import { conditions } from './triggerConditions';
 import { hasTriggerCondition, isInIframe } from './triggerUtils';
 
+class IgnoredScheduledTriggerError extends Error {
+	constructor(message) {
+		super(message);
+		this.name = 'IgnoredScheduledTriggerError';
+	}
+}
+
 class Triggers {
 	/** @property {Triggers} instance*/
 
@@ -68,8 +75,20 @@ class Triggers {
 	}
 
 	async when(id, condition) {
+		const { user } = store.state;
+
 		if (!this._enabled) {
-			return Promise.reject('Triggers disabled');
+			return new IgnoredScheduledTriggerError('Failed to schedule. Triggers disabled.');
+		}
+
+		if (condition.name !== 'after-guest-registration' && user) {
+			throw new IgnoredScheduledTriggerError('Failed to schedule. User already registered.');
+		}
+
+		const record = this._findRecordById(id);
+
+		if (record && record.status === 'scheduled') {
+			throw new IgnoredScheduledTriggerError('Trigger already scheduled. ignoring...');
 		}
 
 		this._updateRecord(id, {
@@ -107,16 +126,25 @@ class Triggers {
 
 		return this.when(id, condition)
 			.then(() => this.fire(id, action, condition))
-			.catch((error) => console.error(`[Livechat Triggers]: ${error}`));
+			.catch((error) => {
+				if (error instanceof IgnoredScheduledTriggerError) {
+					console.warn(`[Livechat Triggers]: ${error}`);
+					return;
+				}
+				console.error(`[Livechat Triggers]: ${error}`);
+			});
 	}
 
 	scheduleAll(triggers) {
 		triggers.map((trigger) => this.schedule(trigger));
 	}
 
+	async processTrigger(id) {
+		this.processTriggers({ force: true, filter: (trigger) => trigger.conditions.some(({ name }) => name === id) });
+	}
+
 	async processTriggers({ force = false, filter = () => true } = {}) {
 		const triggers = this._triggers.filter((trigger) => force || this._isValid(trigger)).filter(filter);
-
 		this.scheduleAll(triggers);
 	}